Refactor Job related classes
