+++ skyve-war/src/main/java/modules/admin/JobSchedule/JobSchedule.xml
@@ -10,7 +10,7 @@
 	<icon16x16RelativeFilePath>icons/document/job16.png</icon16x16RelativeFilePath>
 	<icon32x32RelativeFilePath>icons/document/job32.png</icon32x32RelativeFilePath>
 	<bizKey>
-		return modules.admin.JobSchedule.JobScheduleBizlet.getBizKey(this);
+		return ((JobScheduleExtension)this).bizKey();
 	</bizKey>
 	<attributes>
 		<text name="jobName" required="true">
